+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Metadata field `gcp_pubsub_publish_time_unix` added to `gcp_pubsub` input.
+
 ## 2.13.0 - 2019-08-27
 
 ### Added